Add files via upload
A quick version of high score display, designed for the code to be implemented ultimately in our end screen controller. It is relatively simple, I have left comments on my thought process for certain decisions and improvements I would like to make if I have time when I return from break. The primary thing is since we have no way of tracking player scores just yet I have not yet added the capacity for a player's high score to be added to the list. I left some notes on how this might be achieved with what I have written already so if anyone else wants to take a swing at it they are welcome.